# apps/management-api/openapi.yml
openapi: 3.0.3
info:
  title: Podverse Management API
  description: Comprehensive OpenAPI specification generated from audited route matrix coverage for Podverse management API.
  version: 2.0.0
servers:
  - url: https://management-api.podverse.fm/api/v2
    description: Production Server
  - url: http://localhost:3100/api/v2
    description: Local Development Server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pv_mgmt_auth
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AuthenticationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Authentication failure reason
              example: Unauthorized
            code:
              type: string
              example: unauthorized
    AuthorizationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Authorization failure reason
              example: Forbidden
            code:
              type: string
              example: forbidden
...
    InternalServerErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              example: Internal server error
            code:
              type: string
              example: internal_server_error
...
    BillingCheckoutChannel:
      type: object
      required:
        [id, processor_id, platform, storefront_allowlist, enabled, min_client_version, updated_at]
      properties:
        id: { type: integer }
        processor_id: { type: string, enum: [paypal, apple, google_play, test] }
        platform: { type: string, enum: [web, ios, android, foss] }
        storefront_allowlist: { type: array, items: { type: string } }
        enabled: { type: boolean }
        min_client_version: { type: string, nullable: true }
        updated_at: { type: string, format: date-time }
    BillingCheckoutChannelPatchRequest:
      type: object
      minProperties: 1
      properties:
        enabled: { type: boolean }
        min_client_version: { type: string, nullable: true }
        storefront_allowlist: { type: array, items: { type: string } }
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
  /billing/checkout-channels:
    get:
      operationId: managementGetBillingCheckoutChannels
      tags:
        - billing
      summary: GET /billing/checkout-channels
      description: "Lists every checkout channel. Permission: billing_channels read."
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BillingCheckoutChannel"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /billing/checkout-channels/{id}:
    patch:
      operationId: managementPatchBillingCheckoutChannelId
      tags:
        - billing
      summary: PATCH /billing/checkout-channels/{id}
      description: "Updates enabled, the minimum client version, or the storefront allowlist. Permission: billing_channels update."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingCheckoutChannelPatchRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BillingCheckoutChannel"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /billing/processor-products:
    get:
      operationId: managementGetBillingProcessorProducts
      tags:
        - billing
      summary: GET /billing/processor-products
      description: "Lists processor product mappings. Permission: billing_processor_products read."
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BillingProcessorProduct"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"

    post:
      operationId: managementPostBillingProcessorProducts
      tags:
        - billing
      summary: POST /billing/processor-products
      description: "Maps a processor product id to the Premium product. Permission: billing_processor_products create."
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingProcessorProductCreateRequest"
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BillingProcessorProduct"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /billing/processor-products/{id}:
    patch:
      operationId: managementPatchBillingProcessorProductId
      tags:
        - billing
      summary: PATCH /billing/processor-products/{id}
      description: "Updates is_active or the external ids. Permission: billing_processor_products update."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingProcessorProductPatchRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BillingProcessorProduct"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /billing/accounts/{accountId}:
    get:
      operationId: managementGetBillingAccount
      tags:
        - billing
      summary: GET /billing/accounts/{accountId}
      description: "Subscriptions, transactions, grants, and recent webhook events for one account. Permission: billing_account read."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BillingAccountDetail"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /billing/accounts/{accountId}/resync:
    post:
      operationId: managementPostBillingAccountResync
      tags:
        - billing
      summary: POST /billing/accounts/{accountId}/resync
      description: "Fetches each subscription from its processor and retries the account failed webhook events. Permission: billing_account update."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BillingResyncResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /billing/accounts/{accountId}/grants:
    post:
      operationId: managementPostBillingAccountGrants
      tags:
        - billing
      summary: POST /billing/accounts/{accountId}/grants
      description: "Grants one period of membership. Permission: billing_account create."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingGrantRequest"
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BillingGrantResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
...